feat: add GoString for type schema
