+++ internal/v2/scope/scope.go
@@ -1,0 +1,1 @@
+package scope
