+++ internal/v2/scope/utils.go
@@ -1,0 +1,1 @@
+package scope
